+++ src/components/ui-radio.vue
@@ -1,10 +1,8 @@
 <template>
     <div>
-        <div v-for="(item, index) in items" :key="item.label" class="py-2">
+        <div class="py-2">
             <input
-                :id="`${id}_${index}`"
-                v-model="inputValue"
-                :value="item.value"
+                :id="id"
                 :name="name"
                 class="hidden"
                 type="radio"
@@ -13,12 +11,12 @@
             />
             <ui-label
                 class="cursor-pointer flex py-1 hover:text-blue-1"
-                :for="`${id}_${index}`"
+                :for="id"
             >
                 <div
                     class="bg-blue-5 mr-2 mt-1 w-4 h-4 border border-blue-2 rounded-3xl"
                 ></div>
-                {{ item.label }}
+                <slot />
             </ui-label>
         </div>
     </div>
@@ -32,14 +30,6 @@ export default {
     },
     inheritAttrs: false,
     props: {
-        selected: {
-            type: [String, Number, Boolean],
-            required: true,
-        },
-        items: {
-            type: [Array, String],
-            required: true,
-        },
         id: {
             type: String,
             required: true,
@@ -47,17 +37,6 @@ export default {
         name: {
             type: String,
             default: 'funda_radio_input',
-        },
-    },
-    emits: ['change:value'],
-    computed: {
-        inputValue: {
-            get() {
-                return this.selected;
-            },
-            set(value) {
-                this.$emit('change:value', value);
-            },
         },
     },
 };